[BUGFIX] remove top level side effect from circular module reference that causes race condition dependent on ordering of module load.
